fix performacne issue with concatenation
