+++ src/obfuscationMethods/polygonObfuscation.ts
@@ -1,0 +1,3 @@
+export const dividePolygon = (polygonSvg: SVGElement) => {
+    console.log(polygonSvg);
+}
